18 - Travis CI Support
* Added Travis CI Support.
* Simplified Mocha/Chai Unit Tests
* Changed dry-parser.js to index.js
